+++ Nezabudal/src/main/java/sk/upjs/paz1c/nezabudal/dao/implementations/MysqlCategoryDao.java
@@ -65,6 +65,8 @@ public class MysqlCategoryDao implements CategoryDao {
 
     @Override
     public void delete(Category category) {
+        if (category == null)
+            return;
         String sql = "DELETE FROM category WHERE id = ?";
         jdbcTemplate.update(sql, category.getId());
     }
